+++ CHANGELOG.md
@@ -18,8 +18,10 @@ Please mark all change in change log and use the issue from GitHub
 - \#766 - If partition tag is similar, wrong partition is searched
 - \#771 - Add server build commit info interface
 - \#759 - Put C++ sdk out of milvus/core
+- \#813 - Add push mode for prometheus monitor
 - \#815 - Support MinIO storage
 - \#823 - Support binary vector tanimoto metric
+- \#910 - Change Milvus c++ standard to c++17
 
 ## Improvement
 - \#738 - Use Openblas / lapack from apt install
@@ -29,6 +31,7 @@ Please mark all change in change log and use the issue from GitHub
 - \#848 - Add ready-to-use config files to the Milvus repo for enhanced user experince
 - \#860 - Remove redundant checks in CacheMgr's constructor
 - \#908 - Move "primary_path" and "secondary_path" to storage config
+- \#931 - Remove "collector" from config
 
 ## Task